feat: use iconify universal icon loader
