+++ examples/nuxt/package.json
@@ -12,7 +12,7 @@
     "nuxt": "^2.15.8"
   },
   "devDependencies": {
-    "@iconify/json": "^2.1.12",
+    "@iconify/json": "^2.1.17",
     "@nuxt/types": "^2.15.8",
     "@nuxt/typescript-build": "^2.1.0",
     "@nuxtjs/composition-api": "^0.32.0",
