+++ examples/sveltekit/package.json
@@ -11,10 +11,10 @@
   "devDependencies": {
     "@iconify-json/icon-park": "^1.1.1",
     "@iconify-json/logos": "^1.1.3",
-    "@iconify-json/mdi": "^1.1.1",
+    "@iconify-json/mdi": "^1.1.2",
     "@sveltejs/kit": "next",
     "svelte": "^3.46.4",
-    "svelte-check": "^2.4.5",
+    "svelte-check": "^2.4.6",
     "svelte-preprocess": "^4.10.4",
     "tslib": "^2.3.1",
     "typescript": "^4.6.2",
