+++ examples/sveltekit/svelte.config.js
@@ -7,8 +7,6 @@ const config = {
   // for more information about preprocessors
   preprocess: preprocess(),
   kit: {
-    // hydrate the <div id="svelte"> element in src/app.html
-    // target: '#svelte',
     vite: {
       plugins: [
         Icons({
