+++ examples/vite-react/package.json
@@ -14,8 +14,8 @@
   "devDependencies": {
     "@iconify-json/logos": "^1.1.3",
     "@svgr/core": "^6.2.1",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^17.0.41",
+    "@types/react-dom": "^17.0.14",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "typescript": "^4.6.2",
     "unplugin-icons": "workspace:*",
