+++ examples/vite-solid/package.json
@@ -10,12 +10,12 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@iconify-json/mdi": "^1.1.1",
+    "@iconify-json/logos": "^1.1.3",
     "unplugin-icons": "workspace:*",
     "vite": "^2.8.6",
     "vite-plugin-solid": "^2.2.6"
   },
   "dependencies": {
-    "solid-js": "^1.3.10"
+    "solid-js": "^1.3.13"
   }
 }
