+++ examples/vite-svelte/package.json
@@ -12,12 +12,12 @@
   "devDependencies": {
     "@iconify-json/icon-park": "^1.1.1",
     "@iconify-json/logos": "^1.1.3",
-    "@iconify-json/mdi": "^1.1.1",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.39",
+    "@iconify-json/mdi": "^1.1.2",
+    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.40",
     "@tsconfig/svelte": "^3.0.0",
     "cross-env": "^7.0.3",
     "svelte": "^3.46.4",
-    "svelte-check": "^2.4.5",
+    "svelte-check": "^2.4.6",
     "svelte-preprocess": "^4.10.4",
     "typescript": "^4.6.2",
     "unplugin-icons": "workspace:*",
