+++ examples/vite-vue2/package.json
@@ -8,12 +8,12 @@
     "vue": "^2.6.14"
   },
   "devDependencies": {
-    "@iconify-json/carbon": "^1.1.1",
+    "@iconify-json/carbon": "^1.1.2",
     "@iconify-json/fa-solid": "^1.1.1",
     "@iconify-json/heroicons-outline": "^1.1.1",
-    "@iconify-json/ic": "^1.1.1",
+    "@iconify-json/ic": "^1.1.2",
     "@iconify-json/logos": "^1.1.3",
-    "@iconify-json/mdi": "^1.1.1",
+    "@iconify-json/mdi": "^1.1.2",
     "@iconify-json/mdi-light": "^1.1.1",
     "@iconify-json/noto-v1": "^1.1.1",
     "@iconify-json/ri": "^1.1.1",
