+++ examples/vite-vue3-legacy/package.json
@@ -8,7 +8,7 @@
     "vue": "^3.2.26"
   },
   "devDependencies": {
-    "@iconify/json": "^2.1.12",
+    "@iconify/json": "^2.1.17",
     "@vitejs/plugin-vue": "^2.2.4",
     "@vue/compiler-sfc": "^3.2.31",
     "typescript": "^4.6.2",
