+++ examples/vite-vue3/package.json
@@ -8,21 +8,21 @@
     "vue": "^3.2.26"
   },
   "devDependencies": {
-    "@iconify-json/carbon": "^1.1.1",
+    "@iconify-json/carbon": "^1.1.2",
     "@iconify-json/fa-solid": "^1.1.1",
     "@iconify-json/heroicons-outline": "^1.1.1",
-    "@iconify-json/ic": "^1.1.1",
+    "@iconify-json/ic": "^1.1.2",
     "@iconify-json/icon-park": "^1.1.1",
     "@iconify-json/logos": "^1.1.3",
-    "@iconify-json/mdi": "^1.1.1",
+    "@iconify-json/mdi": "^1.1.2",
     "@iconify-json/mdi-light": "^1.1.1",
     "@iconify-json/noto-v1": "^1.1.1",
     "@iconify-json/ri": "^1.1.1",
-    "@iconify-json/twemoji": "^1.1.1",
+    "@iconify-json/twemoji": "^1.1.2",
     "@vitejs/plugin-vue": "^2.2.4",
     "@vue/compiler-sfc": "^3.2.31",
     "typescript": "^4.6.2",
-    "unconfig": "^0.3.1",
+    "unconfig": "^0.3.2",
     "unplugin-icons": "workspace:*",
     "unplugin-vue-components": "^0.17.21",
     "vite": "^2.8.6"
