+++ examples/vue-cli/package.json
@@ -12,10 +12,10 @@
     "vue": "^2.6.14"
   },
   "devDependencies": {
-    "@iconify/json": "^2.1.12",
-    "@vue/cli-plugin-babel": "^5.0.1",
-    "@vue/cli-plugin-typescript": "^5.0.1",
-    "@vue/cli-service": "^5.0.1",
+    "@iconify/json": "^2.1.17",
+    "@vue/cli-plugin-babel": "^5.0.3",
+    "@vue/cli-plugin-typescript": "^5.0.3",
+    "@vue/cli-service": "^5.0.3",
     "typescript": "^4.6.2",
     "unplugin-icons": "workspace:*",
     "unplugin-vue2-script-setup": "^0.10.0",
