+++ package.json
@@ -115,29 +115,29 @@
   "dependencies": {
     "@antfu/install-pkg": "^0.1.0",
     "@antfu/utils": "^0.5.0",
-    "@iconify/utils": "^1.0.23",
-    "debug": "^4.3.3",
+    "@iconify/utils": "^1.0.26",
+    "debug": "^4.3.4",
     "kolorist": "^1.5.1",
     "local-pkg": "^0.4.1",
     "unplugin": "^0.4.0"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.18.8",
-    "@iconify/json": "^2.1.14",
-    "@iconify/types": "^1.0.12",
+    "@iconify/json": "^2.1.17",
+    "@iconify/types": "^1.0.13",
     "@svgr/core": "^6.2.1",
     "@types/debug": "^4.1.7",
     "@types/node": "^17.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
     "@vue/compiler-sfc": "^3.2.31",
     "bumpp": "^7.1.1",
     "cross-env": "^7.0.3",
     "eslint": "^8.11.0",
     "esno": "^0.14.1",
-    "rollup": "^2.70.0",
+    "rollup": "^2.70.1",
     "tsup": "^5.12.1",
     "typescript": "^4.6.2",
     "vite": "^2.8.6",
-    "vitest": "^0.6.0"
+    "vitest": "^0.6.3"
   }
 }
