+++ src/core/compilers/jsx.ts
@@ -1,5 +1,5 @@
 import { importModule } from 'local-pkg'
-import { camelize } from '../utils'
+import { camelize } from '@iconify/utils/lib/misc/strings'
 import type { Compiler } from './types'
 
 export const JSXCompiler = <Compiler>(async(
